+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/util/PaginationFilterSink.java
@@ -84,4 +84,8 @@ public interface PaginationFilterSink<T> {
 
   public void restrictPaginationByTransferType(T item, TransferType transferType, Consumer<String> errorHandler);
 
+  public void restrictPaginationByTissueOrigin(T item, String origin, Consumer<String> errorHandler);
+
+  public void restrictPaginationByTissueType(T item, String type, Consumer<String> errorHandler);
+
 }