+++ miso-web/src/it/java/uk/ac/bbsrc/tgac/miso/webapp/integrationtest/ListTablesIT.java
@@ -30,14 +30,14 @@ import uk.ac.bbsrc.tgac.miso.webapp.integrationtest.page.element.DataTable;
 
 public class ListTablesIT extends AbstractIT {
 
-  private static final Set<String> samplesColumns = Sets.newHashSet(Columns.SORT, Columns.NAME, Columns.ALIAS, Columns.SAMPLE_CLASS,
-      Columns.TISSUE_ORIGIN, Columns.TISSUE_TYPE, Columns.QC, Columns.LOCATION, Columns.CREATION_DATE, Columns.LAST_MODIFIED);
+  private static final Set<String> samplesColumns = Sets.newHashSet(Columns.SORT, Columns.NAME, Columns.ALIAS, Columns.TISSUE_ORIGIN,
+      Columns.TISSUE_TYPE, Columns.SAMPLE_CLASS, Columns.QC, Columns.LOCATION, Columns.CREATION_DATE, Columns.LAST_MODIFIED);
   private static final Set<String> librariesColumns = Sets.newHashSet(Columns.SORT, Columns.NAME, Columns.ALIAS, Columns.TISSUE_ORIGIN,
-      Columns.TISSUE_TYPE, Columns.QC, Columns.DESIGN, Columns.INDICES, Columns.LOCATION, Columns.VOLUME, Columns.CONCENTRATION,
-      Columns.LAST_MODIFIED);
+      Columns.TISSUE_TYPE, Columns.QC, Columns.DESIGN, Columns.SIZE_BP, Columns.INDICES, Columns.LOCATION, Columns.VOLUME,
+      Columns.CONCENTRATION, Columns.LAST_MODIFIED);
   private static final Set<String> libraryAliquotsColumns = Sets.newHashSet(Columns.SORT, Columns.NAME, Columns.ALIAS,
-      Columns.TISSUE_ORIGIN, Columns.TISSUE_TYPE, Columns.DESIGN, Columns.INDICES, Columns.LOCATION, Columns.VOLUME, Columns.CONCENTRATION,
-      Columns.LAST_MODIFIED);
+      Columns.TISSUE_ORIGIN, Columns.TISSUE_TYPE, Columns.DESIGN, Columns.SIZE_BP, Columns.INDICES, Columns.LOCATION, Columns.VOLUME,
+      Columns.CONCENTRATION, Columns.LAST_MODIFIED);
   private static final Set<String> poolsColumns = Sets.newHashSet(Columns.SORT, Columns.NAME, Columns.ALIAS,
       Columns.DESCRIPTION, Columns.DATE_CREATED, Columns.LIBRARY_ALIQUOTS, Columns.CONCENTRATION, Columns.LOCATION, Columns.AVG_INSERT_SIZE,
       Columns.LAST_MODIFIED);