+++ sqlstore/src/it/java/uk/ac/bbsrc/tgac/miso/persistence/impl/HibernateLibraryAliquotDaoIT.java
@@ -321,6 +321,16 @@ public class HibernateLibraryAliquotDaoIT extends AbstractDAOTest {
     testSearch(PaginationFilter.pool(1L));
   }
 
+  @Test
+  public void testSearchByTissueOrigin() throws Exception {
+    testSearch(PaginationFilter.tissueOrigin("Ly"));
+  }
+
+  @Test
+  public void testSearchByTissueType() throws Exception {
+    testSearch(PaginationFilter.tissueType("P"));
+  }
+
   /**
    * Verifies Hibernate mappings by ensuring that no exception is thrown by a search
    *