+++ sqlstore/src/main/java/uk/ac/bbsrc/tgac/miso/persistence/impl/HibernatePoolableElementViewDao.java
@@ -1,6 +1,7 @@
 package uk.ac.bbsrc.tgac.miso.persistence.impl;
 
 import java.io.IOException;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.Date;
 import java.util.List;
@@ -31,6 +32,8 @@ public class HibernatePoolableElementViewDao implements PoolableElementViewDao, 
   // Make sure these match the HiberateLibraryAliquotDao
   private static final String[] SEARCH_PROPERTIES = new String[] { "aliquotName", "aliquotBarcode", "libraryName", "libraryAlias",
       "libraryDescription" };
+  private final static List<String> STANDARD_ALIASES = Arrays.asList("sample", "sample.hierarchyAttributes",
+      "hierarchyAttributes.tissueOrigin", "hierarchyAttributes.tissueType");
 
   @Autowired
   private SessionFactory sessionFactory;
@@ -89,7 +92,7 @@ public class HibernatePoolableElementViewDao implements PoolableElementViewDao, 
 
   @Override
   public Iterable<String> listAliases() {
-    return Collections.emptyList();
+    return STANDARD_ALIASES;
   }
 
   @Override
@@ -129,6 +132,12 @@ public class HibernatePoolableElementViewDao implements PoolableElementViewDao, 
       return "aliquotNgUsed";
     case "volumeUsed":
       return "aliquotVolumeUsed";
+    case "dnaSize":
+      return "aliquotDnaSize";
+    case "effectiveTissueOriginLabel":
+      return "tissueOrigin.alias";
+    case "effectiveTissueTypeLabel":
+      return "tissueType.alias";
     default:
       return original;
     }
@@ -244,6 +253,16 @@ public class HibernatePoolableElementViewDao implements PoolableElementViewDao, 
     criteria.createAlias("transfers", "transferItem")
         .createAlias("transferItem.transfer", "transfer")
         .add(Restrictions.ilike("transfer.recipient", recipient, MatchMode.ANYWHERE));
+  }
+
+  @Override
+  public void restrictPaginationByTissueOrigin(Criteria criteria, String origin, Consumer<String> errorHandler) {
+    criteria.add(Restrictions.eq("tissueOrigin.alias", origin));
+  }
+
+  @Override
+  public void restrictPaginationByTissueType(Criteria criteria, String type, Consumer<String> errorHandler) {
+    criteria.add(Restrictions.eq("tissueType.alias", type));
   }
 
 }