+++ RELEASE_NOTES.md
@@ -6,6 +6,9 @@ Changes:
     on the Edit Run page
   * Freeze "Parent Alias" column on the bulk Propagate Samples page
   * Allow removing box from freezer on Edit Box page
+  * Added size (bp) column to Library and Library Aliquot lists
+  * Allow searching and sorting by tissue origin and tissue type (detailed sample)
+  * Allow sorting by volume and concentration on Library and Library Aliquot lists
   * Fixed detailed sample pages loading
   * Fixed error updating transfer items
   * Fixed receipt validation errors when date of receipt is not specified